Refactor user-permissions file structure
