+++ src/domain/config/repositories/UserMonitoringConfigRepository.ts
@@ -1,4 +1,4 @@
-import { UsersOptions } from "domain/entities/UserMonitoring";
+import { UsersOptions } from "domain/entities/user-monitoring/UserMonitoring";
 
 export interface UserMonitoringConfigRepository {
     get(): Promise<UsersOptions>;
