+++ src/domain/entities/user-monitoring/UserGroup.ts
@@ -1,0 +1,4 @@
+export interface UserGroup {
+    id: string;
+    name: string;
+}
